+++ ui/theme/components/icons/index.js
@@ -82,4 +82,5 @@ export * from "./UserFill";
 export * from "./ParentGroupIcon";
 export * from "./ClockIcon";
 export * from "./CrossError";
+export * from "./Support";
 export * from "./AlertRounded";
